+++ concurrency/Control/Concurrent/Classy/QSemN.hs
@@ -20,6 +20,7 @@ import           Control.Concurrent.Classy.MVar
 import           Control.Monad.Catch            (mask_, onException,
                                                  uninterruptibleMask_)
 import           Control.Monad.Conc.Class       (MonadConc)
+import           Control.Monad.Fail             (MonadFail)
 import           Data.Maybe
 
 -- | 'QSemN' is a quantity semaphore in which the resource is aqcuired
@@ -39,7 +40,7 @@ newtype QSemN m = QSemN (MVar m (Int, [(Int, MVar m ())], [(Int, MVar m ())]))
 --  The initial quantity must be at least 0.
 --
 -- @since 1.0.0.0
-newQSemN :: MonadConc m => Int -> m (QSemN m)
+newQSemN :: (MonadConc m, MonadFail m) => Int -> m (QSemN m)
 newQSemN initial
   | initial < 0 = fail "newQSemN: Initial quantity must be non-negative"
   | otherwise   = QSemN <$> newMVar (initial, [], [])